+++ src/main/java/calculator/exception/StringAddCalculatorException.java
@@ -1,0 +1,8 @@
+package calculator.exception;
+
+public class StringAddCalculatorException extends RuntimeException {
+
+    public StringAddCalculatorException(String message) {
+        super(message);
+    }
+}
