+++ src/main/java/lotto/exception/LottoException.java
@@ -1,0 +1,8 @@
+package lotto.exception;
+
+public class LottoException extends RuntimeException {
+
+    public LottoException(String msg) {
+        super(msg);
+    }
+}
