--- src/main/java/lotto/step1/exception/StringAddCalculatorException.java
@@ -1,8 +1,0 @@
-package lotto.step1.exception;
-
-public class StringAddCalculatorException extends RuntimeException {
-
-    public StringAddCalculatorException(String message) {
-        super(message);
-    }
-}
