--- src/main/java/lotto/step2/domain/Lotto.java
@@ -1,52 +1,0 @@
-package lotto.step2.domain;
-
-import lotto.step2.exception.LottoException;
-
-import java.util.Objects;
-import java.util.Set;
-import java.util.stream.Stream;
-
-public class Lotto {
-
-    private static final String LOTTO_NUMBERS_MADE_OF_SIX_NUMBERS = "로또는 6개의 숫자로 이루어져야 합니다.";
-    private static final int LOTTO_MADE_OF_NUMBERS = 6;
-    private Set<Integer> numbers;
-
-    private Lotto(final Set<Integer> numbers) {
-        validation(numbers);
-        this.numbers = numbers;
-    }
-
-    public static Lotto of(final Set<Integer> numbers) {
-        return new Lotto(numbers);
-    }
-
-    public Stream<Integer> stream() {
-        return numbers.stream();
-    }
-
-    private void validation(final Set<Integer> numbers) {
-        if (numbers.size() != LOTTO_MADE_OF_NUMBERS) {
-            throw new LottoException(LOTTO_NUMBERS_MADE_OF_SIX_NUMBERS);
-        }
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        Lotto lotto = (Lotto) o;
-        return Objects.equals(numbers, lotto.numbers);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(numbers);
-    }
-}
