--- src/main/java/lotto/step2/exception/LottoException.java
@@ -1,8 +1,0 @@
-package lotto.step2.exception;
-
-public class LottoException extends RuntimeException {
-
-    public LottoException(String msg) {
-        super(msg);
-    }
-}
