--- src/test/java/lotto/step2/domain/LottoTest.java
@@ -1,40 +1,0 @@
-package lotto.step2.domain;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-
-import java.util.Arrays;
-import java.util.Set;
-import java.util.TreeSet;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class LottoTest {
-
-    @DisplayName("로또는 6개의 숫자로 이루어짐")
-    @ParameterizedTest
-    @MethodSource("provideLottos")
-    void lotto_숫자6개구성(int[] input, int[] expected) {
-        Set<Integer> lottoSet = new TreeSet<>(Arrays.stream(input)
-                .boxed()
-                .collect(Collectors.toSet()));
-        Set<Integer> expectedSet = new TreeSet<>(Arrays.stream(expected)
-                .boxed()
-                .collect(Collectors.toSet()));
-
-        Lotto lotto = Lotto.of(lottoSet);
-        assertThat(lotto).isEqualTo(Lotto.of(expectedSet));
-    }
-
-    private static Stream<Arguments> provideLottos() {
-        return Stream.of(
-                Arguments.of(new int[]{1, 3, 5, 7, 9, 11}, new int[]{1, 3, 5, 7, 9, 11}),
-                Arguments.of(new int[]{5, 10, 13, 17, 29, 31}, new int[]{5, 10, 13, 17, 29, 31}),
-                Arguments.of(new int[]{21, 33, 34, 37, 40, 45}, new int[]{21, 33, 34, 37, 40, 45})
-        );
-    }
-}
